Refactor front-end entrypoint bootstrap
